backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
beautifulsoup4==4.12.3
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
anthropic==0.18.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
lxml==5.1.0
python-dateutil==2.8.2
psycopg2-binary==2.9.9
